+++ include/can_plugins2/can_utils.hpp
@@ -1,0 +1,65 @@
+#pragma once
+
+#include <array>
+#include <can_plugins2/msg/frame.hpp>
+
+namespace can_utils
+{
+
+  inline constexpr int CAN_MTU = 8;
+
+  enum class Endian{
+    little,
+    big
+  };
+
+  template <typename T>
+  inline void can_unpack(const std::array<uint8_t, CAN_MTU> &buf, T &data)
+  {
+    memcpy(&data, buf.data(), sizeof(T));
+  }
+
+  template <typename T>
+  inline void can_pack(std::array<uint8_t, CAN_MTU> &buf, const T data)
+  {
+    memcpy(buf.data(), &data, sizeof(T));
+  }
+
+  template <typename T>
+  inline std::unique_ptr<can_plugins2::msg::Frame> generate_frame(const uint16_t id, const T data)
+  {
+    can_plugins2::msg::Frame frame;
+    frame.id = id;
+    frame.is_rtr = false;
+    frame.is_extended = false;
+    frame.is_error = false;
+
+    frame.dlc = sizeof(T);
+    frame.data.fill(0);
+
+    can_pack<T>(frame.data, data);
+
+    return make_unique<can_plugins2::msg::Frame>(frame);
+  }
+
+  template <typename T>
+  inline std::unique_ptr<can_plugins2::msg::Frame> generate_frame(const uint16_t id, const T data, const Endian endian)
+  {
+    can_plugins2::msg::Frame frame;
+    frame.id = id;
+    frame.is_rtr = false;
+    frame.is_extended = false;
+    frame.is_error = false;
+
+    frame.dlc = sizeof(T);
+    frame.data.fill(0);
+
+    can_pack<T>(frame.data, data);
+    if(Endian == big){
+      std::reverse(frame.data.begin(), frame.data.end());
+    }
+
+    return make_unique<can_plugins2::msg::Frame>(frame);
+  }
+
+} // namespace can_utils
