+++ include/can_plugins2/safe_array.hpp
